requirements.txt
requests
lxml
Pillow
tqdm
customtkinter
CTkMessagebox